+++ Modular Gameplay Overhaul/Modules/Tools/Configs/ScytheConfig.cs
@@ -23,15 +23,15 @@ public sealed class ScytheConfig
 
     /// <summary>Gets a value indicating whether to harvest crops.</summary>
     [JsonProperty]
-    public bool HarvestCrops { get; internal set; } = true;
+    public bool HarvestCrops { get; internal set; } = false;
 
     /// <summary>Gets a value indicating whether to harvest flowers.</summary>
     [JsonProperty]
-    public bool HarvestFlowers { get; internal set; } = true;
+    public bool HarvestFlowers { get; internal set; } = false;
 
     /// <summary>Gets a value indicating whether to harvest spring onions.</summary>
     [JsonProperty]
-    public bool HarvestSpringOnions { get; internal set; } = true;
+    public bool HarvestSpringOnions { get; internal set; } = false;
 
     /// <summary>Gets a value indicating whether the scythes can be enchanted with Haymaker.</summary>
     [JsonProperty]